refactor(ui): extract RemoteImagePasteController
